task detail layout restructuring
