+++ lib/helpers/format_date.dart
@@ -1,0 +1,7 @@
+import 'package:intl/intl.dart';
+
+class FormatDate {
+  static String toDmmmmyyyy(DateTime date) {
+    return '${DateFormat.d().format(date)} ${DateFormat.MMMM('ru').format(date)} ${DateFormat.y().format(date)}';
+  }
+}
