+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'package:todo/presentation/screens/all_tasks.dart';
-import 'package:todo/presentation/screens/task_detail.dart';
+import 'package:todo/presentation/screens/task_detail/task_detail.dart';
 
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:todo/repositories/tasks_repository.dart';
