+++ pubspec.yaml
@@ -28,6 +28,7 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  intl: ^0.17.0
   uuid: ^3.0.7
   flutter_bloc: ^8.1.3
   rxdart: ^0.27.7
